Move feed utils to separate file
